+++ README.md
@@ -1,6 +1,12 @@
-#Laratest
+# Laratest
 
 Tools for test your Laravel app.
+
+Publish config file:
+
+```php
+php artisan vendor:publish --provider="Jandrodev\Laratest\LaratestServiceProvider" --tag="config"
+```
 
 ---
